Create ExamResult.java
You have been provided with an incomplete Java class named ExamResult. This class has a method isPass(int marks) which needs to be completed.

The isPass(int marks) method should take an integer value representing the marks obtained by a student in an exam and determine whether the student has passed or failed the exam.

In this case, a student is considered to have passed the exam if they have scored more than 50 marks. If the student's marks are 50 or less, they are considered to have failed.